+++ serialisation/src/test/java/org/ehrbase/openehr/sdk/serialisation/dto/OldDtoModelTest.java
@@ -38,11 +38,10 @@ import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.old.coronaanamnes
 import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.old.coronaanamnesecomposition.definition.VorhandenDefiningcode;
 import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.old.testalltypesenv1composition.TestAllTypesEnV1Composition;
 import org.ehrbase.openehr.sdk.serialisation.RMDataFormat;
+import org.ehrbase.openehr.sdk.serialisation.TestDataTemplateProvider;
 import org.ehrbase.openehr.sdk.serialisation.flatencoding.FlatFormat;
 import org.ehrbase.openehr.sdk.serialisation.flatencoding.FlatJasonProvider;
 import org.ehrbase.openehr.sdk.serialisation.jsonencoding.ArchieObjectMapperProvider;
-import org.ehrbase.openehr.sdk.serialisation.jsonencoding.CanonicalJson;
-import org.ehrbase.openehr.sdk.serialisation.templateprovider.TestDataTemplateProvider;
 import org.ehrbase.openehr.sdk.test_data.composition.CompositionTestDataCanonicalJson;
 import org.ehrbase.openehr.sdk.test_data.composition.CompositionTestDataSimSDTJson;
 import org.junit.Test;
@@ -117,7 +116,7 @@ public class OldDtoModelTest {
 
     @Test
     public void testFlattenCorona() throws IOException {
-        Composition composition = new CanonicalJson()
+        Composition composition = RMDataFormat.canonicalJSON()
                 .unmarshal(
                         IOUtils.toString(CompositionTestDataCanonicalJson.CORONA.getStream(), StandardCharsets.UTF_8),
                         Composition.class);
@@ -132,7 +131,7 @@ public class OldDtoModelTest {
 
     @Test
     public void testUnflattenCorona() throws IOException {
-        Composition expected = new CanonicalJson()
+        Composition expected = RMDataFormat.canonicalJSON()
                 .unmarshal(
                         IOUtils.toString(CompositionTestDataCanonicalJson.CORONA.getStream(), StandardCharsets.UTF_8),
                         Composition.class);